+++ plugins/woocommerce-admin/packages/components/src/text-control/docs/example.js
@@ -1,0 +1,27 @@
+/** @format */
+/**
+ * Internal dependencies
+ */
+import { TextControl } from '@woocommerce/components';
+
+/**
+ * External dependencies
+ */
+import { withState } from '@wordpress/compose';
+
+export default withState( {
+	value: '',
+} )( ( { setState, value } ) => {
+	return (
+		<div>
+			<TextControl
+				name="text-control"
+				label="Enter text here"
+				onChange={ newValue => setState( { value: newValue } ) }
+				value={ value }
+			/>
+			<br />
+			<TextControl label="Disabled field" disabled value="" />
+		</div>
+	);
+} );